+++ src/blender-addon/util/layout.py
@@ -1,0 +1,9 @@
+import textwrap
+
+def label_multiline(context, text, parent):
+  chars = int(context.region.width / 7)   # 7 pix on 1 character
+  wrapper = textwrap.TextWrapper(width=chars)
+  text_lines = wrapper.wrap(text=text)
+  for text_line in text_lines:
+    parent.label(text=text_line)
+